Avoid using separate libraries for subparts of Core
In particular for Elements, Layout and FontEngineDefault. Defining these as interface libraries cause issues during export. I believe we would have to export them as separate targets, but that doesn't relly make sense from a client standpoint and causes new issues with exported source files and dependencies. It seems to be a lot less troublesome to use a single CMake library and set properties and attach sources to that directly.

Setup runtime output directory from root, and other minor tweaks.